refactor : order exception 정리
