--- src/main/java/com/ybe/tr1ll1on/domain/order/exception/OrderNotFoundException.java
@@ -1,8 +1,0 @@
-package com.ybe.tr1ll1on.domain.order.exception;
-
-import com.ybe.tr1ll1on.global.exception.ExceptionCode;
-import com.ybe.tr1ll1on.global.exception.TrillionException;
-
-public class OrderNotFoundException extends TrillionException {
-    public OrderNotFoundException (ExceptionCode errorCode) { super(errorCode); }
-}
